negative value set no limit for package lifetime and buffer size

+++ server/src/configure/configure.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <iostream>
+#include <limits>
 
 struct configure{
     int port;
     float package_lifetime;
-    int buffer_size;
+    unsigned int buffer_size;
 };
 
 void set_port_conf(int port);
@@ -13,5 +14,5 @@ void set_package_lifetime_conf(float package_lifetime);
 void set_buffer_size_conf(int buffer_size);
 
 float get_package_lifetime_conf();
-float get_port_conf();
-float get_buffer_size_conf();
+int get_port_conf();
+unsigned int get_buffer_size_conf();
